+++ Modules/StudyProgramme/test/ilStudyProgrammeLPTest.php
@@ -46,7 +46,9 @@ class ilStudyProgrammeLPTest extends PHPUnit_Framework_TestCase {
 	}
 	
 	protected function tearDown() {
-		$this->root->delete();
+		if ($this->root) {
+			$this->root->delete();
+		}
 	}
 	
 	protected function newUser() {
