+++ Modules/StudyProgramme/test/ilStudyProgrammeProgressCalculationTest.php
@@ -36,7 +36,9 @@ class ilStudyProgrammeProgressCalculationTest extends PHPUnit_Framework_TestCase
 	}
 	
 	protected function tearDown() {
-		$this->root->delete();
+		if ($this->root) {
+			$this->root->delete();
+		}
 	}
 	
 	protected function newUser() {
